Feat: extend indentation (#79)
* refactor: extract indentation regex to customisable variable

This allows users to configure this regex to match their own style.

* feat: allow white space in front semicolon for indentation

Allow white space in between the last object and semicolon on a line. This way
statements following a line such as `?s ?p ?o ;` are intended the same way as
when the semicolon directly follows the object.

* feat: allow comments at the end of a line for indentation

Indent statements following a line where the last statement ends with a
semicolon, irrelevant of whether that lines ends with a comment. This way a
statement following a line such as `?s ?p ?o; # comment` is indented to same way
as when the comment would not be there.

* feat: indent to object after comma

When the statement on the previous line ends with a comma, indent the current
line to the its object.

* fix: put OR alternatives in a group

Put the alternatives for the OR expression into a group. Otherwise the first
alternative matches any statement with a comma, including function calls where
the arguments are separated by a comma.